Add birthday querying commands
